Add window parameter to parameters
The window is the number of images with which an output frame will be
made of.
Suppose a sequence of frames:
 __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __
|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|

if window == 3, the first 3 frames will be summed up, divided by three
and saved as the first frame of output.

 __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __
|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|
 |  / /
 | //'
 __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __ __
|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|__|

The window size also determines the maximum number of threads. Even if
the number of threads is passed as parameter, it can't exceed
(n_frames - window)/2

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>